+++ test/plausible/imported/google_analytics4_test.exs
@@ -150,6 +150,87 @@ defmodule Plausible.Imported.GoogleAnalytics4Test do
       assert_active_visitors(site_import)
     end
 
+    test "handles empty response payload gracefully", %{user: user, site: site} do
+      future = DateTime.add(DateTime.utc_now(), 3600, :second)
+
+      empty_custom_events = %{
+        "reports" => [
+          %{
+            "dimensionHeaders" => [
+              %{"name" => "date"},
+              %{"name" => "eventName"},
+              %{"name" => "linkUrl"}
+            ],
+            "kind" => "analyticsData#runReport",
+            "metadata" => %{"currencyCode" => "USD", "timeZone" => "Etc/GMT"},
+            "metricHeaders" => [
+              %{"name" => "totalUsers", "type" => "TYPE_INTEGER"},
+              %{"name" => "eventCount", "type" => "TYPE_INTEGER"}
+            ]
+          }
+        ]
+      }
+
+      {:ok, job} =
+        Plausible.Imported.GoogleAnalytics4.new_import(
+          site,
+          user,
+          label: "properties/123456",
+          property: "properties/123456",
+          start_date: ~D[2024-01-01],
+          end_date: ~D[2024-01-31],
+          access_token: "redacted_access_token",
+          refresh_token: "redacted_refresh_token",
+          token_expires_at: DateTime.to_iso8601(future)
+        )
+
+      site_import = Plausible.Imported.get_import(site, job.args.import_id)
+
+      opts = job |> Repo.reload!() |> Map.get(:args) |> GoogleAnalytics4.parse_args()
+
+      opts = Keyword.put(opts, :flush_interval_ms, 10)
+
+      for report <- Enum.take(@full_report_mock, 4) do
+        expect(Plausible.HTTPClient.Mock, :post, fn _url, headers, _body, _opts ->
+          assert [{"Authorization", "Bearer redacted_access_token"}] == headers
+          {:ok, %Finch.Response{status: 200, body: report}}
+        end)
+      end
+
+      expect(Plausible.HTTPClient.Mock, :post, fn _url, headers, _body, _opts ->
+        assert [{"Authorization", "Bearer redacted_access_token"}] == headers
+        {:ok, %Finch.Response{status: 200, body: empty_custom_events}}
+      end)
+
+      for report <- Enum.drop(@full_report_mock, 5) do
+        expect(Plausible.HTTPClient.Mock, :post, fn _url, headers, _body, _opts ->
+          assert [{"Authorization", "Bearer redacted_access_token"}] == headers
+          {:ok, %Finch.Response{status: 200, body: report}}
+        end)
+      end
+
+      assert :ok = GoogleAnalytics4.import_data(site_import, opts)
+
+      Enum.each(Plausible.Imported.tables(), fn table ->
+        count =
+          case table do
+            "imported_sources" -> 210
+            "imported_visitors" -> 31
+            "imported_pages" -> 3340
+            "imported_entry_pages" -> 2934
+            "imported_exit_pages" -> 0
+            "imported_locations" -> 2291
+            "imported_devices" -> 93
+            "imported_browsers" -> 233
+            "imported_operating_systems" -> 1068
+            "imported_custom_events" -> 0
+          end
+
+        query = from(imported in table, where: imported.site_id == ^site.id)
+        assert await_clickhouse_count(query, count)
+      end)
+    end
+
     test "handles rate limiting gracefully", %{user: user, site: site} do
       future = DateTime.add(DateTime.utc_now(), 3600, :second)